Moved RedisConnector into connector package
