+++ src/main/scala/play/api/cache/redis/connector/RedisConnector.scala
@@ -1,4 +1,4 @@
-package play.api.cache.redis
+package play.api.cache.redis.connector
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{ExecutionContext, Future}
