+++ src/main/scala/play/api/cache/redis/package.scala
@@ -8,4 +8,5 @@ package object redis extends AnyRef with util.Expiration {
   type SynchronousResult[ A ]  = A
   type AsynchronousResult[ A ] = scala.concurrent.Future[ A ]
 
+  type RedisConnector = connector.RedisConnector
 }
